Pong Game - Python
Pong is a game written in Python using the PyGame library. It is a classic two-player game where players control a paddle and hit the ball back and forth. The first player who misses the ball loses! Making a Pong project in Python means writing code to run the game.